+++ CHANGELOG.md
@@ -80,6 +80,7 @@ IMPROVEMENTS / BUG FIXES:
   - Box adding doesn't use `/tmp` anymore which can avoid some cross-device
     copy issues. [GH-1199]
   - Vagrant works properly in folders with strange characters. [GH-1223]
+  - Vagrant properly handles "paused" VirtualBox machines. [GH-1184]
 
 ## 1.0.6 (December 21, 2012)